python-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.31.0
pymupdf==1.24.11
langchain==0.3.0
langchain-anthropic==0.3.0
python-multipart==0.0.12
pydantic==2.9.0
httpx==0.27.0
python-dotenv==1.0.1
